+++ lib/run_loop/core.rb
@@ -10,6 +10,8 @@ module RunLoop
 
   module Core
 
+    include RunLoop::Regex
+
     START_DELIMITER = "OUTPUT_JSON:\n"
     END_DELIMITER="\nEND_OUTPUT"
 
@@ -256,7 +258,7 @@ Please update your sources to pass an instance of RunLoop::Xcode))
             }
       merged_options = options.merge(discovered_options)
 
-      if self.simulator_target?(merged_options, sim_control)
+      if self.simulator_target?(merged_options)
         self.expect_compatible_simulator_architecture(merged_options, sim_control)
         self.prepare_simulator(merged_options, sim_control)
       end
@@ -353,7 +355,9 @@ Please update your sources to pass an instance of RunLoop::Xcode))
     #   `self.udid_and_bundle_for_launcher`
     #
     # @see {Core::RunLoop.udid_and_bundle_for_launcher}
-    def self.simulator_target?(run_options, sim_control = RunLoop::SimControl.new)
+    #
+    # @todo sim_control argument is no longer necessary and can be removed.
+    def self.simulator_target?(run_options, sim_control=nil)
       value = run_options[:device_target]
 
       # match the behavior of udid_and_bundle_for_launcher
@@ -362,15 +366,7 @@ Please update your sources to pass an instance of RunLoop::Xcode))
       # support for 'simulator' and Xcode >= 5.1 device targets
       return true if value.downcase.include?('simulator')
 
-      # if Xcode < 6.0, we are done
-      return false if not sim_control.xcode_version_gte_6?
-
-      # support for Xcode >= 6 simulator udids
-      return true if sim_control.sim_udid? value
-
-      # support for Xcode >= 6 'named simulators'
-      sims = sim_control.simulators.each
-      sims.find_index { |device| device.name == value } != nil
+      value[DEVICE_UDID_REGEX, 0] == nil
     end
 
     # Extracts the value of :inject_dylib from options Hash.
@@ -442,7 +438,7 @@ Please update your sources to pass an instance of RunLoop::Xcode))
         end
         udid = device_target
 
-        unless self.simulator_target?(options, sim_control)
+        unless self.simulator_target?(options)
           bundle_dir_or_bundle_id = options[:bundle_id] if options[:bundle_id]
         end
       else